Extract Back Office root utils
