+++ apps/back-office/src/app/_utils/getFullNLAddress.ts
@@ -1,18 +1,11 @@
 import type { MeldingOutput } from '~/apiClientProxy'
 
-type ShortAddressInput = Pick<MeldingOutput, 'street' | 'house_number' | 'house_number_addition'>
+import { getShortNLAddress } from './getShortNLAddress'
+
 type FullAddressInput = Pick<
   MeldingOutput,
   'street' | 'house_number' | 'house_number_addition' | 'postal_code' | 'city'
 >
-
-export const getShortNLAddress = ({ house_number, house_number_addition, street }: ShortAddressInput) => {
-  if (!street || !house_number) return undefined
-
-  const addition = house_number_addition ? house_number_addition : ''
-
-  return `${street} ${house_number}${addition}`
-}
 
 export const getFullNLAddress = ({
   city,
